requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.7
future==0.18.2
gunicorn==20.0.4
idna==2.9
line-bot-sdk==1.16.0
psycopg2==2.8.5
pytz==2020.1
requests==2.23.0
sqlparse==0.3.1
static3==0.7.0
urllib3==1.25.9
